fix: 修复其中读取不确定关系时错误的类别转换
	修改：     xml_operator.py

+++ xml_operator.py
@@ -310,7 +310,7 @@ def xml_unsure_relational_partial_retrieval(input_xml):
 			text = relation.text
 			_relation = {
 				'type':  relation.get('type'),
-				'confidence': int(relation.get('confidence')),
+				'confidence': float(relation.get('confidence')),
 				'source': relation.get('source'),
 				'target': int(relation.find('target').text)
 			}
